srcs/requirements/service_user_handler/django/conf/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.13.0
whitenoise==6.6.0
